rename recipes/ -> modDefinitions/
