+++ src/modDefinitions/recipesHooks.ts
@@ -17,7 +17,7 @@
 
 import { type ModDefinition } from "@/types/modDefinitionTypes";
 import { useDispatch, useSelector } from "react-redux";
-import { selectAllRecipes } from "@/recipes/recipesSelectors";
+import { selectAllRecipes } from "@/modDefinitions/recipesSelectors";
 import { useCallback, useEffect } from "react";
 import { recipesActions } from "./recipesSlice";
 import { type RegistryId } from "@/types/registryTypes";
