+++ src/pageEditor/store.ts
@@ -38,8 +38,8 @@ import { authSlice, persistAuthConfig } from "@/auth/authSlice";
 import analysisSlice from "@/analysis/analysisSlice";
 import pageEditorAnalysisManager from "./analysisManager";
 import { tabStateSlice } from "@/pageEditor/tabState/tabStateSlice";
-import { recipesSlice } from "@/recipes/recipesSlice";
-import { recipesMiddleware } from "@/recipes/recipesListenerMiddleware";
+import { recipesSlice } from "@/modDefinitions/recipesSlice";
+import { recipesMiddleware } from "@/modDefinitions/recipesListenerMiddleware";
 import { type StorageInterface } from "@/store/StorageInterface";
 import {
   persistSessionChangesConfig,
